DataTranslation: Fix crash
just noticed this crash...

when fConfigView gets deleted by selecting a translator it
deletes its child fInfoText but we were leaving the pointer
alone.

Afterwords if you changed the panel text color in Appearance
a message gets sent that checks to see if the fInfoView
pointer is NULL, and since it isn't, procedes to dereference
the pointer and *boom* the app crashes.

Fix this by setting the fInfoText pointer to NULL when fConfig
view gets deleted. That way when you change the panel text
color it doesn't attempt to dereference the stale pointer
and everything works as it should.