Mandelbrot: Scrap the old app, initial workings of a new one.
Based upon a HTML5 fractal engine by @kerwizzy (relicensed as MIT and copyright
assigned to Haiku, Inc. with permission from the author). Hacky and a mess,
but it works, and has the potentiality to be much nicer than the previous
Mandelbrot app.

Already supports rendering & zooming in, but has hacks all over the place.