LegacyPackageInstaller: Correct previous commit.
The test in the previous commit had a totally flawed invocation of strcmp().
I probably shouldn't attempt to write C code without consulting the manual.

Even if it had worked, it would have worked on too many cases, including
some volume paths. So I reordered the if statement and added it in there.