XHCI: Remove the whitelist and enable by default for all controllers.
Discussed with kallisti5 and PulkoMandy before committing. So far, the
only bugs since korli's recent changes either are minor ones that don't
appear to affect functionality (e.g. #12860), or so bad that nothing works
(so, just as good as if the controller was disabled.) PulkoMandy reports
that his controller works but wasn't whitelisted, as have some other users,
so I'm enabling it by default.

If there turn out to be too many broken devices, we can revert this. Or only
revert it in the beta branch, if needed.