LICENSE: Rename to License.md, and remove all licenses but the MIT.
This file was introduced with the rationale that bots attempt to scrape
GitHub repositories for license data. GitHub now has an API for this,
and since all its sniffer rules fail because this file is not identical
to any known license, it just marks it "Other" and includes the full text
anyway.

Since this is the case, nobody else wants to look at all the licenses
in a single monolithic file, so we just put the explanatory message
and a reference to the directory containing all the licenses at the top
of the file, and include only the MIT license.

Also, GitHub renders Markdown nicer than it does plain text, so use that
(which matches what we already do for READMEs.)