Bullet: Remove license file and AboutSystem credit.
We aren't really using Bullet itself; there are only two classes
(Quaternion and Vector3) which are from Bullet, and even they were
heavily modified. Furthermore, the Bullet license is really just the Zlib
license, which the Bullet source repository confirms (and you can see
by looking at its text.) It does not require attribution in documentation,
but only in the source code, and we aren't even using a substantial part
of it.