add WIP "netconsole" KDL add-on
Similar to Linux netconsole, this should send all debug output over UDP
to the broadcast address. It can be listened to with netcat or similar tools.
This could be an easy way to extract KDL info from a crashing machine
when no serial ports are available.

Does not work yet, help welcome.