Keymap: Refactor _LoadSettings()
Instead of filling out the path into a BString and using it in the calling method,
call _SetKeyboardLayout() directly using the path found.

Also use a plain const char* array instead of BString because we don't need
the added functionality.

The return value now reflects whether or not the window frame was loaded.

Check the return value on the caller side and only use the resulting window
frame if it is B_OK, otherwise CenterOnScreen().