Keymap: Reinitialize ref after creating Keymap dir
In the case that the Keymap directory does not exist and has to
be created, reinitialize ref to the newly created directory.

Thanks Stippi for your help

There is still a bug present where if the Keymap directory is a
symlink to another directory, the File Panel does not follow that
link and instead opens to your home directory. More on this later.