Keymap: Follow symlink
get_ref_for_path() doesn't resolve symlinks.

Construct a BEntry from the path passing 'true' for the traverse
argument. That BEntry is then be passed to BDirectory's constructor.

The error was hidden by the fact that the code didn't bother to error
check the result of get_ref_for_path(), which would have indicated that
the passed in path was a link.

Thanks Rene for your help.

if the entry fails to init, we are probably screwed, but, try to get
the path using get_ref_for_path in that case. Worst case scenario it
fails as well and the FilePanel points to your home directory.