PS/2: fix dead code.
hrev50506 extended the range of acceptable X and Y movement for PS/2
mouse. However, the value is sent on 1 byte (+ sign) so the check would
now always be valid.

Remove the check completely, we now rely only on bit 3 of the first byte
to be set (and not missing packets) to remain in sync with the mouse.

Thanks to Marcus Overhagen for reviewing the changes.