Updated mail provider info for gmx.de
Activated SSL encrption for POP3 and changed server name for SMTP to
mail.gmx.net. The first change is vital to get it to work, the latter
seems to be only cosmetic.

I still get a "503 Bad sequence of commands" when sending a mail, but
it is sent nonetheless. See ticket: #12960.