MBR: Mark CHS value as invalid, not unused
As reported in #6301, some BIOSes rely on these byte to detect an unused
partition entry (set to 0) from an used one (set to some other value),
and only later notice that the CHS is invalid and LBA should be used.

Apply the patch that Ingo had attached to the ticket back then.

As far as I can tell, the protective MBR in GPT only happens to use the
same way to mark the partition as used (in case you try to plug a GPT
drive to a BIOS that does only CHS). So, using the same value should not
lead to the MBR being identified as a protective MBR by GPT aware
systems (that is detected using the partition type and partition size).

Fixes #6301.