libgnuregex: fix debug build.
Add missing "printchar.c", which is only used in DEBUG mode. The
function was recovered from original regex sources.

Note that the gnu regexp library is deprecated, and so is its
replacement, gnu rx. GNU suggests using the regex implementation from
the glibc, but that one isn't as portable.

Thanks to Andrew Lindesay for the investigations and initial patch!

Fixes #12952.