3rdparty/qtcreator: Totally new version of the create_project_file script.
Qt Creator now has a "generic project" mode, in which it just acts
as an auto-completing code editor. I tried using it on the entire
Haiku project at once, but it's just too much for Qt Creator to handle.
So instead, I created a script which generates project files for
any given directory in the tree, as well as sets up the proper include
directories. The project files themselves are .gitignore'd; use the script
to create them.

Works on Haiku. Did not test on Linux with a crosstools setup; but
it should work there too.