intel_extreme: set image size on both pipes.
This was hardcoded to pipe A on my laptop, but it's better to set both
pipes so it work in more cases.

Eventually I should see why my SandyBridge laptop is failing to move its
LVDS panel to pipe B...

This fixes miqlas Thinpad X200s and possibly some other machines. Thanks
for letting me experiment with the hardware!