LegacyPackageInstaller: Ignore files which do not end in ".pdb".
The LegacyPackageInstaller's data folder is (and has always been)
"~/config/packages", which is of course the same directory that Haiku's
(relatively) new package management system uses. We don't want the
"administrative" folder and any HPKGs in that directory to show up in
the "installed legacy packages" list, so ignore them.