AVFormatReader: take the frame size into account
In ffmpeg, an audio frame may be made of multiple samples. We don't
really support that in our plug-in, and assume the frame rate and sample
rate are the same thing.

Adjust the reported frame count by multiplying ffmpeg's frame count by
its frame size (which is the number of samples per frame).

This fixes audio not working or stopping after a short time in
MediaPlayer, which uses the frame count this way.