Deskbar: Avoid leaking leaf bitmap's memory
In the vector case we are creating a bitmap, in the non-vector
case we are grabbing the bitmap from a resource.

The vector case we have to create a bitmap to rasterize.
The non-vector case we just grab the bitmap pointer.

We have to create new bitmaps from the vector icon at
different sizes, when we do that we were leaking the memory
occupied by the previous icon.

So make and use SetIcon() setter method which deletes the old
bitmap before setting the new one. In the non-vector case this
means we need to make a copy of the bitmap we grab from the
resources that we can safely delete.

Also delete bitmap on destructor.

For all other TBarMenuTitle's (meaning the team menu item),
we are grabbing a bitmap from resource, so don't delete the
bitmap as it is owned by resource set, we just use the pointer.