HaikuDepot: Improve icon download handling performance
Previously each icon would launch an independent HTTP request to
pull down the HVIF icon data.  This change means that the data
will be pulled down in bulk across all packages as a .tgz and
will then be kept in a cache locally.  The client-server logic
will use standard "If-Modified-Since" headers to check for
updates each time the HaikuDepot desktop application starts up.
This arrangement will bring down the HVIF as well as bitmap
icons and use the best representation it can.

Additionally, it is possible from a command-line option to log
HTTP traffic verbosely and it is also possible to use an "-h"
flag to display help on command-line arguments.

The code-structure around this change also anticipates some
future extensions to handle other client-server improvements.

Fixes #11804