Added square Haiku logos
These are made from the sticker "haiku-sticker-3.5x2".
Quick preview at http://img.ctrlv.in/img/17/02/27/58b425e23a6fb.png

A "Haiku" text layer and a 1-pixel frame are on separate layers,
that are set invisible. Not sure if we should always show the "Haiku".
Looks better without in my opinion, but depending on where it'll be
used, showing the text may be desired...

When used for small logos, like  64px, simple scaling might not cut
it, and the yellow and orange "hills" need to be made wider.