ffmpeg: give a little hint about the file type.
ffmpeg, especially older versions, is not very good at detecting file
types and relies on us giving it a file extension for a better guess.

Use our MIME sniffing to have a first try at detecting the file we are
about to play, and pass this information to ffmpeg.

Fixes #13150.