Do not fail when the display mode is sanitized
The goal of sanitize_display_mode is to create a mode that the hardware
can use. The code was doing that, but if any change was required to the
initial mode, it threw the result away.

Now accept the sanitized mode, and log it for debugging purposes.

This should fix #12927