GPT: lower our identify score if only one header is present.
GPT should have its partition table both at the start and end of the
disk. Our GPT partitionning system is able to cope with just one of the
two being available. However this has the unwanted effect that when
erasing only the start of the disk, it is still recognized with its old
GPT scheme.

To avoid this, lower the score of GPT if it finds only one of the two
headers, allowing MBR and BFS to get more priority in that case.

Should fix #12290, #13434.